Challen's Initial Visualizations
A collection of a few exploratory visualizations on the housing dataset. The main finding is near the bottom, where I aggregated square footage into bins and plotted a bar chart to illustrate how increasing the number of bedrooms increases price, showing how it would be a worthy variable to include in our predictive model.